requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==4.2.8
django-cockroachdb==4.2
django-phonenumber-field==8.0.0
google-auth==2.32.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
idna==3.7
oauthlib==3.2.2
pdfkit==1.0.0
phonenumbers==8.13.40
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-decouple==3.8
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
